refactor(bench): split result aggregation helpers
